+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added progress bar component
+- Added base button component
+- Added list component
+- Added possibility to change between app mode and storybook app mode
 
 ## [1.1.1] - 2023-10-31